requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-jwt-extended==4.3.1
python-dotenv==0.19.0
datetime==4.3
python-dateutil==2.8.2